Move Occurrence parsers into metadata package
