+++ pkg/kritis/metadata/grafeas/grafeas.go
@@ -126,7 +126,7 @@ func (c Client) Vulnerabilities(containerImage string) ([]metadata.Vulnerability
 	}
 	var vulnz []metadata.Vulnerability
 	for _, occ := range occs {
-		if v := util.GetVulnerabilityFromOccurrence(occ); v != nil {
+		if v := metadata.GetVulnerabilityFromOccurrence(occ); v != nil {
 			vulnz = append(vulnz, *v)
 		}
 	}
@@ -142,7 +142,7 @@ func (c Client) Attestations(containerImage string, aa *kritisv1beta1.Attestatio
 		return nil, err
 	}
 	for _, occ := range occs {
-		ra, err := util.GetRawAttestationsFromOccurrence(occ)
+		ra, err := metadata.GetRawAttestationsFromOccurrence(occ)
 		if err != nil {
 			return nil, err
 		}
